Change in DownloadFileService and remove the dependence of this service on Window

+++ CrossPlatformDownloadManager.Data/Services/DownloadQueueService/DownloadQueueService.cs
@@ -1,8 +1,11 @@
 using System.Collections.ObjectModel;
 using AutoMapper;
+using Avalonia.Controls;
 using CrossPlatformDownloadManager.Data.Models;
+using CrossPlatformDownloadManager.Data.Services.DownloadFileService;
 using CrossPlatformDownloadManager.Data.Services.UnitOfWork;
 using CrossPlatformDownloadManager.Data.ViewModels;
+using CrossPlatformDownloadManager.Data.ViewModels.CustomEventArgs;
 using CrossPlatformDownloadManager.Utils;
 using PropertyChanged;
 
@@ -15,6 +18,8 @@ public class DownloadQueueService : IDownloadQueueService
 
     private readonly IUnitOfWork _unitOfWork;
     private readonly IMapper _mapper;
+    private readonly IDownloadFileService _downloadFileService;
+    private readonly List<DownloadQueueTaskViewModel> _downloadQueueTasks;
 
     #endregion
 
@@ -30,10 +35,13 @@ public class DownloadQueueService : IDownloadQueueService
 
     #endregion
 
-    public DownloadQueueService(IUnitOfWork unitOfWork, IMapper mapper)
+    public DownloadQueueService(IUnitOfWork unitOfWork, IMapper mapper, IDownloadFileService downloadFileService)
     {
         _unitOfWork = unitOfWork;
         _mapper = mapper;
+        _downloadFileService = downloadFileService;
+
+        _downloadQueueTasks = [];
 
         DownloadQueues = [];
     }
@@ -45,7 +53,7 @@ public class DownloadQueueService : IDownloadQueueService
 
         var downloadQueues = await _unitOfWork
             .DownloadQueueRepository
-            .GetAllAsync();
+            .GetAllAsync(includeProperties: ["DownloadFiles"]);
 
         var primaryKeys = downloadQueues
             .Select(df => df.Id)
@@ -56,14 +64,14 @@ public class DownloadQueueService : IDownloadQueueService
             .ToList();
 
         foreach (var downloadQueue in exceptDownloadQueues)
-            await DeleteDownloadQueueAsync(downloadQueue);
+            await DeleteDownloadQueueAsync(downloadQueue: downloadQueue, reloadData: false);
 
         var downloadQueueViewModels = _mapper.Map<List<DownloadQueueViewModel>>(downloadQueues);
         foreach (var downloadQueue in downloadQueueViewModels)
         {
             var oldDownloadQueue = DownloadQueues.FirstOrDefault(dq => dq.Id == downloadQueue.Id);
             if (oldDownloadQueue != null)
-                UpdateDownloadQueueViewModel(oldDownloadQueue, downloadQueue);
+                oldDownloadQueue.UpdateViewModel(downloadQueue, nameof(oldDownloadQueue.Id));
             else
                 DownloadQueues.Add(downloadQueue);
         }
@@ -71,8 +79,11 @@ public class DownloadQueueService : IDownloadQueueService
         DataChanged?.Invoke(this, EventArgs.Empty);
     }
 
-    public async Task AddNewDownloadQueueAsync(DownloadQueue downloadQueue, bool reloadData = true)
+    public async Task AddNewDownloadQueueAsync(DownloadQueue? downloadQueue, bool reloadData = true)
     {
+        if (downloadQueue == null)
+            return;
+
         await _unitOfWork.DownloadQueueRepository.AddAsync(downloadQueue);
         await _unitOfWork.SaveAsync();
 
@@ -80,62 +91,78 @@ public class DownloadQueueService : IDownloadQueueService
             await LoadDownloadQueuesAsync();
     }
 
-    public async Task DeleteDownloadQueueAsync(DownloadQueueViewModel downloadQueueViewModel)
+    public async Task DeleteDownloadQueueAsync(DownloadQueueViewModel? downloadQueue, bool reloadData = true)
     {
+        if (downloadQueue == null)
+            return;
+
         var downloadQueueInDb = await _unitOfWork
             .DownloadQueueRepository
-            .GetAsync(dq => dq.Id == downloadQueueViewModel.Id);
+            .GetAsync(dq => dq.Id == downloadQueue.Id);
 
         if (downloadQueueInDb == null)
             return;
-        
+
         _unitOfWork.DownloadQueueRepository.Delete(downloadQueueInDb);
         await _unitOfWork.SaveAsync();
-        await LoadDownloadQueuesAsync();
+
+        if (reloadData)
+            await LoadDownloadQueuesAsync();
     }
 
-    public async Task UpdateDownloadQueueAsync(DownloadQueue downloadQueue)
+    public async Task UpdateDownloadQueueAsync(DownloadQueue? downloadQueue)
     {
+        if (downloadQueue == null)
+            return;
+
         await _unitOfWork.DownloadQueueRepository.UpdateAsync(downloadQueue);
         await _unitOfWork.SaveAsync();
         await LoadDownloadQueuesAsync();
     }
 
-    public async Task StartDownloadQueueAsync(DownloadQueueViewModel downloadQueueView)
+    public void StartDownloadQueue(DownloadQueueViewModel? downloadQueue)
     {
+        if (downloadQueue == null || downloadQueue.DownloadFiles.Count == 0)
+            return;
+        
+        downloadQueue.IsRunning = true;
+        ContinueDownloadQueue(downloadQueue);
     }
 
-    public async Task StopDownloadQueueAsync(DownloadQueueViewModel downloadQueueViewModel)
+    public async Task StopDownloadQueueAsync(DownloadQueueViewModel? downloadQueue)
     {
+        if (downloadQueue == null)
+            return;
+
+        var downloadFiles = downloadQueue
+            .DownloadFiles
+            .Where(df => df.IsDownloading)
+            .ToList();
+
+        if (downloadFiles.Count == 0)
+            return;
+
+        foreach (var downloadFile in downloadFiles)
+            await _downloadFileService.StopDownloadFileAsync(downloadFile);
+
+        downloadQueue.IsRunning = false;
     }
 
-    public async Task AddDownloadFileToDownloadQueueAsync(DownloadQueueViewModel downloadQueueViewModel,
-        DownloadFileViewModel downloadFileViewModel)
+    public async Task AddDownloadFileToDownloadQueueAsync(DownloadQueueViewModel? downloadQueue,
+        DownloadFileViewModel? downloadFile)
     {
+        if (downloadQueue == null || downloadFile == null)
+            return;
     }
 
-    public async Task RemoveDownloadFileFromDownloadQueueAsync(DownloadQueueViewModel downloadQueueViewModel,
-        DownloadFileViewModel downloadFileViewModel)
+    public async Task RemoveDownloadFileFromDownloadQueueAsync(DownloadQueueViewModel? downloadQueue,
+        DownloadFileViewModel? downloadFile)
     {
+        if (downloadQueue == null || downloadFile == null)
+            return;
     }
 
     #region Helpers
-
-    private void UpdateDownloadQueueViewModel(DownloadQueueViewModel? oldDownloadQueue,
-        DownloadQueueViewModel? newDownloadQueue)
-    {
-        if (oldDownloadQueue == null || newDownloadQueue == null)
-            return;
-
-        var properties = newDownloadQueue
-            .GetType()
-            .GetProperties()
-            .Where(p => !p.Name.Equals("Id", StringComparison.OrdinalIgnoreCase) && p.CanWrite)
-            .ToList();
-
-        foreach (var property in properties)
-            property.SetValue(oldDownloadQueue, property.GetValue(newDownloadQueue));
-    }
 
     private async Task AddDefaultDownloadQueueAsync()
     {
@@ -155,6 +182,55 @@ public class DownloadQueueService : IDownloadQueueService
         };
 
         await AddNewDownloadQueueAsync(downloadQueue, reloadData: false);
+    }
+
+    private void ContinueDownloadQueue(DownloadQueueViewModel? downloadQueue)
+    {
+        if (downloadQueue == null || downloadQueue.DownloadFiles.Count == 0)
+            return;
+
+        var downloadFiles = downloadQueue
+            .DownloadFiles
+            .Where(df => !df.IsDownloading)
+            .OrderBy(df => df.DownloadQueuePriority)
+            .ToList();
+
+        var taskIndex = 0;
+        while (_downloadQueueTasks.Count < downloadQueue.DownloadCountAtSameTime)
+        {
+            if (taskIndex == downloadFiles.Count)
+                break;
+
+            var downloadFile = downloadFiles[taskIndex];
+            downloadFile.DownloadFinished += DownloadFileOnDownloadFinished;
+
+            _downloadFileService.StartDownloadFileAsync(downloadFile);
+            _downloadQueueTasks.Add(new DownloadQueueTaskViewModel
+            {
+                Key = downloadFile.Id,
+                DownloadFile = downloadFile,
+                DownloadQueueId = downloadQueue.Id,
+            });
+
+            taskIndex++;
+        }
+    }
+
+    private async void DownloadFileOnDownloadFinished(object? sender, DownloadFileEventArgs e)
+    {
+        var task = _downloadQueueTasks.FirstOrDefault(t => t.Key == e.Id);
+        if (task == null)
+            return;
+
+        task.DownloadFile!.DownloadFinished -= DownloadFileOnDownloadFinished;
+
+        task.DownloadFile!.DownloadQueueId = null;
+        await _downloadFileService.UpdateDownloadFileAsync(task.DownloadFile!);
+
+        var downloadQueue = DownloadQueues.FirstOrDefault(dq => dq.Id == task.DownloadQueueId);
+        ContinueDownloadQueue(downloadQueue);
+
+        _downloadQueueTasks.Remove(task);
     }
 
     #endregion
